+++ DataDefinitions/CommodityMarketQuote.cs
@@ -3,6 +3,7 @@ using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
 using System.Runtime.Serialization;
+using Utilities;
 
 namespace EddiDataDefinitions
 {
@@ -61,15 +62,15 @@ namespace EddiDataDefinitions
         }
 
         // Per-station information
-        public int? buyprice { get; set; }
-        public int? stock { get; set; }
+        public int buyprice { get; set; }
+        public int stock { get; set; }
 
         // StockBracket can contain the values 0, 1, 2, 3 or "" (yes, really) so needs to be dynamic
         // VB: it can be an optional enum no prob
         public dynamic stockbracket { get; set; }
 
-        public int? sellprice { get; set; }
-        public int? demand { get; set; }
+        public int sellprice { get; set; }
+        public int demand { get; set; }
 
         // DemandBracket can contain the values 0, 1, 2, 3 or "" (yes, really) so needs to be dynamic
         // VB: same type as stockbracket above
@@ -82,18 +83,43 @@ namespace EddiDataDefinitions
         [Obsolete("Please use localizedName or InvariantName")]
         public string category => definition?.category.localizedName;
         public int? avgprice => definition?.avgprice;
-        public bool? rare => definition?.rare;
+        public bool rare => definition?.rare ?? false;
 
         public CommodityMarketQuote(CommodityDefinition definition)
         {
             this.definition = definition;
-            buyprice = null;
-            stock = null;
             stockbracket = "";
-            sellprice = null;
-            demand = null;
             demandbracket = "";
             StatusFlags = new List<string>();
+        }
+
+        public static CommodityMarketQuote FromCapiJson(JObject capiJSON)
+        {
+            CommodityDefinition commodityDef = CommodityDefinition.CommodityDefinitionFromEliteID((long)capiJSON["id"]);
+            if (commodityDef == null || (string)capiJSON["name"] != commodityDef.edname)
+            {
+                if (commodityDef.edname != "Drones")
+                {
+                    // Unknown commodity; report the full object so that we can update the definitions
+                    Logging.Info("Commodity definition error: " + (string)capiJSON["name"], JsonConvert.SerializeObject(capiJSON));
+                }
+                return null;
+            }
+            CommodityMarketQuote quote = new CommodityMarketQuote(commodityDef);
+            quote.buyprice = (int)capiJSON["buyPrice"];
+            quote.stock = (int)capiJSON["stock"];
+            quote.stockbracket = (int)capiJSON["stockBracket"];
+            quote.sellprice = (int)capiJSON["sellPrice"];
+            quote.demand = (int)capiJSON["demand"];
+            quote.demandbracket = (int)capiJSON["demandBracket"];
+
+            List<string> StatusFlags = new List<string>();
+            foreach (dynamic statusFlag in capiJSON["statusFlags"])
+            {
+                StatusFlags.Add((string)statusFlag);
+            }
+            quote.StatusFlags = StatusFlags;
+            return quote;
         }
     }
 }