+++ EDDNResponder/EDDNEconomy.cs
@@ -1,8 +1,16 @@
-﻿namespace EDDNResponder
+﻿using EddiDataDefinitions;
+
+namespace EDDNResponder
 {
     class EDDNEconomy
     {
         public string name;
-        public decimal proportion;
+        public decimal proportion = 0M;
+
+        public EDDNEconomy(CompanionAppEconomy companionAppEconomy)
+        {
+            name = companionAppEconomy.name;
+            proportion = companionAppEconomy.proportion;
+        }
     }
 }