+++ EDDI/ChangeLog.md
@@ -2,6 +2,12 @@
 
 Full details of the variables available for each noted event, and VoiceAttack integrations, are available in the individual [event pages](https://github.com/EDCD/EDDI/wiki/Events).
 
+### 3.0.2-b1
+  * Speech responder
+    * Added `Cargo scoop` event
+    * Added `Landing gear` event
+    * Added `Lights` event
+
 ### 3.0.1-rc4
   * Core
     * Removed `insurance excess` from the user interface and the `insurance` property from the top level `commander` variable. FDev now gives us rebuy values directly, so this is no longer needed.